Refactor: one script to install all applications
Add the 'install.sh' script which installs all applications in the 'root/'
and 'local/' folders. The old 'install-local.sh' and 'install-root.sh'
scripts has been adjusted and moved to the 'scripts/' directory.
The installation scripts to run now are selected by the SCRIPTS variable
in the 'to_install.sh' file.

Also create a script to check if the selected installation scripts
exist (scripts/are_valid.sh) and to add execute permission to
the installation scripts (scripts/exec_permission.sh).