Refactor: post installation instructions
The installation scripts no longer saves the post installation messages
in a 'post-install.txt' file. These instructions are in markdown files
inside the 'post_install/' directory.

Add a script that lists the files (and their path) with the
post-installation instructions: 'scripts/post-install.sh'. This script
is called by the 'install.sh' script after the entire installation is
finished.