Feat: Alacritty and Tmux terminfo
Add a terminfo configuration file for Alacritty:
'local/alacritty.terminfo'.

All the Tmux terminfo configurations is now in the 'local/tmux.terminfo' file.
There is a configuration in this file to integrate Tmux with alacritty.
Its name is 'tmux-alacritty'. You need to set it in you '~/.tmux.conf' with the
'default-terminal' option.